requirements.txt
cycler==0.10.0
Django==2.0.5
django-filter==1.1.0
django-filters==0.2.1
djangorestframework==3.8.2
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.2
numpy==1.14.3
Pillow==5.1.0
py-expression-eval==0.3.4
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
six==1.11.0
